backend/requirements.txt
anyio==3.6.2
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
fastapi==0.95.0
filelock==3.11.0
fonttools==4.39.3
gunicorn==20.1.0
h11==0.14.0
idna==3.4
importlib-resources==5.12.0
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.1
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.0
Pillow==9.5.0
psutil==5.9.4
pydantic==1.10.7
pyparsing==3.0.9
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
requests==2.28.2
scipy==1.10.1
seaborn==0.12.2
sentry-sdk==1.19.1
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
sympy==1.11.1
thop==0.1.1.post2209072238
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
ultralytics==8.0.75
urllib3==1.26.15
uvicorn==0.21.1
wincertstore==0.2
zipp==3.15.0
